scrapling-search-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
httpx==0.28.1
python-dotenv==1.0.1
lxml==5.3.0
pydantic-settings==2.6.1
scrapling>=0.2.0
curl_cffi>=0.7.0
browserforge>=1.3.0
